refactor reducer store
