+++ client/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom'
 import {Provider} from 'react-redux'
 import {Router} from 'react-router-dom'
 import history from './history'
-import store from './store'
+import store from './reducers/store'
 import App from './app'
 
 // establishes socket connection
