+++ client/reducers/reducers.js
@@ -1,0 +1,5 @@
+import user from './user';
+
+export default {
+	user,
+}
