+++ client/reducers/store.js
@@ -2,13 +2,17 @@ import {createStore, combineReducers, applyMiddleware} from 'redux'
 import createLogger from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
-import user from './user'
+import reducerList from './reducers';
 
-const reducer = combineReducers({user})
+const reducer = combineReducers({
+	...reducerList
+})
+
 const middleware = composeWithDevTools(applyMiddleware(
   thunkMiddleware,
   createLogger({collapsed: true})
 ))
+
 const store = createStore(reducer, middleware)
 
 export default store
